Extract a container for the GalleryViewThumbnail
